requirements.txt
alembic==1.16.5
androguard==4.1.3
annotated-types==0.7.0
anyio==4.9.0
apkInspector==1.3.5
apkutils==2.0.2
asn1crypto==1.5.1
asttokens==3.0.0
banal==1.0.6
beautifulsoup4==4.13.5
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
cryptography==45.0.5
cvzone==1.6.1
cycler==0.12.1
dataset==1.6.2
DateTime==5.5
decorator==5.2.1
dlib==20.0.0
dnspython==2.7.0
email-validator==2.3.0
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
firebase_admin==7.1.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.59.2
frida==17.2.17
gcloud==0.17.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
google-search==1.1.1
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
gTTS==2.5.4
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jws==0.1.3
kiwisolver==1.4.9
loguru==0.7.3
lxml==6.0.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.1
mutf8==1.0.6
mysql-connector-python==9.4.0
networkx==3.5
numpy==2.2.6
oauth2client==3.0.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
parso==0.8.5
pillow==11.3.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.4.3
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pyelftools==0.27
pyftype==1.2.4
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
Pyrebase==3.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jwt==2.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==0.7.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
sentry-sdk==2.35.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==1.4.54
stack-data==0.6.3
starlette==0.47.3
threadpoolctl==3.6.0
traitlets==5.14.3
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
zope.interface==7.2
gunicorn